requirements.txt
# ─────────────────── Web Framework ───────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ─────────────────── Database ─────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
geoalchemy2==0.14.7
alembic==1.13.1

# ─────────────────── Auth / Security ─────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ─────────────────── Config ───────────────────────────
pydantic-settings==2.2.1

# ─────────────────── File Storage ────────────────────
cloudinary==1.40.0

# ─────────────────── Testing ─────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ─────────────────── Utilities ───────────────────────
python-dotenv==1.0.1
